touching up division of modules in xnat unittests
